+++ coinkiri-application/src/main/kotlin/com/coinkiri/application/service/auth/KakaoAuthService.kt
@@ -1,7 +1,8 @@
-package com.coinkiri.application
+package com.coinkiri.application.service.auth
 
 import com.coinkiri.application.port.`in`.usecase.AuthUseCase
 import com.coinkiri.application.port.out.oauth2.OAuthApiCaller
+import com.coinkiri.application.service.member.MemberService
 import org.springframework.stereotype.Service
 
 @Service
