+++ coinkiri-application/src/main/kotlin/com/coinkiri/application/service/auth/dto/KakaoProfileResponse.kt
@@ -1,4 +1,4 @@
-package com.coinkiri.application
+package com.coinkiri.application.service.auth.dto
 
 data class KakaoProfileResponse(
     val id: String
